divide syntax tree nodes into categories
